requirements.txt
django>=4.0.0<=4.1.0
Pillow
python-dateutil